Fix the use of safe in binary_to_term for cluster
If you have distributed nodes that hold different parts
of your application, which means that the atom table is
different between the nodes are going to have a problem when
you use `:erlang.binary_to_term` on it because it doesn't have
the atoms from the other nodes. So we are getting:
`** (ArgumentError) argument error` when we are running tests
which pointed us to this problem. I know currently eventstore
isn't workin in a distributed way but this still needs to be
addressed. If you are worried about safe encoding then there are
other non lossy encoding you can use.