Added error/3 and FailureContext to Process Managers
Added a struct to be carried on to error handling callbacks on process
managers that contains more metadata.
What we specially added is the state that the process manager would be
on in case of a not failed command execution.

Also, now we have a place to add more data in the future without
breaking the api (error/3 will probably be good for a long time)