Fixes the typespec for command dispatch
When running Dialyxir on a project that dispatches commands with either the `include_execution_result: true` or `include_aggregate_version: true` options set, an error would be encountered due to the incomplete typespec on the `Commanded.Commands.Dispatcher.dispatch/1` function. This change simply updates the documentation and the typespec to match the actual implementation. It is a nonfunctional change.