Add `hibernate_after` option to application config
The `hibernate_after` option is used by the Commanded application's top-level supervisor when starting its children. This determines when these `GenServer` processes hibernate after a period of inactivity.

The default values is `:infinity` which inactivity hibernation is disabled. You can configure the value when defining or starting the Commanded application.