added compile-time config option to globally set aggregate lifespan
The default behavior of "aggregate processes stay around forever" can
present difficulty in a production environment.  For any system that
creates a steady stream of new aggregates, memory creeps up slowly until
exhausted.

It is currently possible to override this on a per-router basis, but
that only creates a situation where forgetting to specify it one place
creates an even more subtle "memory leak".

This commit allows changing the default globally at compile-time.  It
does not allow changing it at run-time, as this seemed like it would
cause unpredictable behavior and possibly negatively impact performance.

Testing this is difficult since it's a compile-time change.  As a
compromise, I set the value to the default explicitly in the testing
config.  While this doesn't fully test the functionality, it does at
least work-out the code.